requirements.txt
Django==1.8.3
Markdown==2.6.2
django-filter==0.10.0
djangorestframework==3.2.1
requests==2.8.1